requirements/local.txt
django==5.2.6
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
django-extra-views==0.16.0